put middleware into lib
